Optimize Dockerfile (#91)
* Restructure Dockerfile

+++ Jenkinsfile
@@ -189,8 +189,14 @@ pipeline {
                   }
                 } finally {
                   library("tailor-meta@${params.tailor_meta}")
+                  try {
+                    if (fileExists(".")) {
+                    deleteDir()
+                    }
+                  } catch (e) {
+                    println e
+                  }
                   cleanDocker()
-                  deleteDir()
                 }
               }}]
             }
